FashionItemEditForm, ImageUploadButton 컴포넌트 수정
1. FashionItemEditForm 컴포넌트
- editFashionItemWithoutImage 메소드를 따로 만들어서
이미지를 수정하지 않을 경우에는 이미지 없이 수정 요청을 해도 문제가 없다는 것을 알 수 있도록 함
- 패션 아이템을 지우기 전에 물어보도록 함

2. ImageUploadButton 컴포넌트
- 이미지 파일을 첨부하지 않았을 경우, 이미지 파일이 첨부되지 않았다는 텍스트를 렌더링하도록 함

+++ src/components/closet/fashionItemEditForm.jsx
@@ -63,16 +63,27 @@ const FashionItemEditForm = ({
     })
   }
 
+  const editFashionItemWithoutImage = async () => {
+    await editFashionItem({
+      fashionItemId: initialFashionItem.id,
+      season,
+      category,
+      color,
+    })
+  }
+
   const handleFashionItemEditButtonClick = async () => {
     try {
-      let uploadedImageId
-
       if (imageFiles) {
         const res = await uploadImage(imageFiles)
-        uploadedImageId = res.data[0].id
+        const uploadedImageId = res.data[0].id
+        await editFashionItemWithImage(uploadedImageId)
       }
 
-      await editFashionItemWithImage(uploadedImageId)
+      if (!imageFiles) {
+        await editFashionItemWithoutImage()
+      }
+
       afterEditFashionItem()
     } catch (error) {
       console.error(error)
@@ -80,6 +91,13 @@ const FashionItemEditForm = ({
   }
 
   const handleFashionItemDeleteButtonClick = async () => {
+    const isFashionItemDelete =
+      window.confirm('패션 아이템을 삭제하시겠습니까?')
+
+    if (!isFashionItemDelete) {
+      return
+    }
+
     try {
       await deleteFashionItem(initialFashionItem.id)
       afterDeleteFashionItem()
